+++ CHANGES.md
@@ -6,8 +6,8 @@
     - make sure `profileNumber` is not reused;
     - make sure the `listen` and `vpnProtoPorts` combinations do not overlap 
       between profiles
-    - make sure there is no overlap in IPv4 ranges (`range` option) assigned to
-      VPN profiles (for now shows warning if there is a problem)
+    - make sure there is no overlap in IP ranges (`range` / `range6` option) 
+      assigned to VPN profiles (for now shows warning if there is a problem)
     - make sure `range` is `/29` or lower per OpenVPN process 
       (OpenVPN limitation)
     - make sure `range6` is `/112` or lower per OpenVPN process